create CaseEventReceiver class to replace part of the GenericEventReceiverImpl class
